src/requirements.txt
click==8.0.3
requests==2.26.0
mypy==0.910
autopep8==1.6.0
semver==2.13.0
packaging==21.0
wheel==0.37.1
setuptools==58.0.4
pyyaml==6.0